Feature/decomposition lu

+++ inc/Matrix.hpp
@@ -10,9 +10,9 @@ class Matrix
 {
 private:
     int _rows, _columns;
-    std::vector<std::vector<long long>> *_table = nullptr; // It's basically the fundamental structure of a Matrix
+    std::vector<std::vector<long double>> *_table = nullptr; // It's basically the fundamental structure of a Matrix
 public:
-    Matrix(int rows, int columns, std::vector<std::vector<long long>> table);
+    Matrix(int rows, int columns, std::vector<std::vector<long double>> table);
     Matrix(const Matrix &); // Deep Copy
     Matrix(int dimension);  // Identity matrix (n x n)
     ~Matrix();
@@ -21,16 +21,16 @@ public:
     void showMatrix();
     int getRows();
     int getColumns();
-    std::vector<std::vector<long long>> getTable();
+    std::vector<std::vector<long double>> getTable();
 
     Matrix &operator=(const Matrix &);
     Matrix operator+(const Matrix &);
     Matrix operator-(const Matrix &);
     Matrix operator*(const Matrix &); // Matrix product
 
-    Matrix operator+(long long scalarNum);
-    Matrix operator-(long long scalarNum);
-    Matrix operator*(long long scalarNum); // Matrix product with scalar number
+    Matrix operator+(long double scalarNum);
+    Matrix operator-(long double scalarNum);
+    Matrix operator*(long double scalarNum); // Matrix product with scalar number
 };
 
 #endif
